+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
   * Improved function specs.
   * Improved generated docs.
-  * Added option to avoid using generated docs.
+  * Added option to avoid automatically generated docs.
+  * Added function for OS environment variable generation for Unix, Releases
+    and Windows.
 
 # Changelog for 2.0.4